Add files via upload
Air pollution monitoring is the systematic observation and analysis of the presence and concentration of pollutants in the air. Using specialized sensors and technology, this process measures contaminants such as particulate matter, nitrogen dioxide, sulfur dioxide, ozone, and carbon monoxide. The gathered data provides insights into the quality of the air, identifies sources of pollution, and assesses the potential health and environmental risks. Monitoring air pollution is crucial for understanding the impact of human activities, implementing regulatory measures, and developing strategies to mitigate pollution and enhance overall air quality. This continuous monitoring helps communities, governments, and organizations make informed decisions to protect public health and the environment.